// web/src/lib/generated/PushDeviceResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PushDevice } from "./PushDevice";

/**
 * Answer of the routes that create or change one device.
 */
export type PushDeviceResponse = { 
/**
 * The device as it is now.
 */
device: PushDevice, };